Revert "perf: move resolver CPU work to worker thread"
This reverts commit 2c4e9cc26307f641c4d8c4311630d6bc1c967be8.
