+++ exec/commands/package.json
@@ -34,6 +34,7 @@
   },
   "dependencies": {
     "@pnpm/bins.resolver": "workspace:*",
+    "@pnpm/cache.metadata": "workspace:*",
     "@pnpm/catalogs.resolver": "workspace:*",
     "@pnpm/cli.command": "workspace:*",
     "@pnpm/cli.common-cli-options-help": "workspace:*",
