+++ exec/commands/tsconfig.json
@@ -19,6 +19,9 @@
       "path": "../../bins/resolver"
     },
     {
+      "path": "../../cache/metadata"
+    },
+    {
       "path": "../../catalogs/resolver"
     },
     {
