+++ store/connection-manager/tsconfig.json
@@ -10,6 +10,9 @@
   ],
   "references": [
     {
+      "path": "../../cache/metadata"
+    },
+    {
       "path": "../../cli/meta"
     },
     {
