+++ worker/package.json
@@ -34,15 +34,11 @@
   },
   "dependencies": {
     "@pnpm/building.pkg-requires-build": "workspace:*",
-    "@pnpm/cache.metadata": "workspace:*",
     "@pnpm/crypto.integrity": "workspace:*",
     "@pnpm/error": "workspace:*",
     "@pnpm/fs.graceful-fs": "workspace:*",
     "@pnpm/fs.hard-link-dir": "workspace:*",
     "@pnpm/fs.symlink-dependency": "workspace:*",
-    "@pnpm/resolving.registry.pkg-metadata-filter": "workspace:*",
-    "@pnpm/resolving.registry.types": "workspace:*",
-    "@pnpm/resolving.resolver-base": "workspace:*",
     "@pnpm/store.cafs": "workspace:*",
     "@pnpm/store.cafs-types": "workspace:*",
     "@pnpm/store.create-cafs-store": "workspace:*",
@@ -50,7 +46,6 @@
     "@rushstack/worker-pool": "catalog:",
     "is-windows": "catalog:",
     "p-limit": "catalog:",
-    "ramda": "catalog:",
     "semver": "catalog:"
   },
   "peerDependencies": {
@@ -61,7 +56,6 @@
     "@pnpm/types": "workspace:*",
     "@pnpm/worker": "workspace:*",
     "@types/is-windows": "catalog:",
-    "@types/ramda": "catalog:",
     "@types/semver": "catalog:"
   },
   "engines": {
