+++ worker/tsconfig.json
@@ -13,9 +13,6 @@
       "path": "../building/pkg-requires-build"
     },
     {
-      "path": "../cache/metadata"
-    },
-    {
       "path": "../core/error"
     },
     {
@@ -35,15 +32,6 @@
     },
     {
       "path": "../fs/symlink-dependency"
-    },
-    {
-      "path": "../resolving/registry/pkg-metadata-filter"
-    },
-    {
-      "path": "../resolving/registry/types"
-    },
-    {
-      "path": "../resolving/resolver-base"
     },
     {
       "path": "../store/cafs"
